chore: Switch to manual APK signing for release builds
Switched from using `build-signed-apk` action to manual APK signing using the Android Gra
dle plugin.
This change allows for more control over the build process and eliminates the need for a third-party action.

The keystore is now written to a temporary file and used during the build process.
The temporary file is then deleted to ensure security.